backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
anthropic==0.43.0
openai==1.59.3
python-multipart==0.0.20
pymupdf==1.25.3
python-docx==1.1.2
tiktoken==0.8.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
anyio==4.8.0
httpx==0.28.1
